requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
line-bot-sdk==3.7.0
pillow==10.1.0
aiohttp==3.9.1 